Improve decompressor compatibility with minitar and Gem::Package::TarReader
Both of these accept io-like objects, which makes it possible to
decompress archives without loading them into memory, as illustrated here:

  https://github.com/halostatue/minitar/tree/3acdc06d182cbd2205ec1acd1619f3d011cf36ca#synopsis

I've resorted to patching Bzip2::FFI::Reader at runtime with a couple of
methods needed to allow implementing a wrapper which can work with
the aforementioned libraries - but would like to upstream these changes
as they could be useful and are minimally invasive.

Firstly `eof?` is needed by both Gem::Package::TarReader and minitar,
the implementation is straightforward as the concept of whether the
output stream has reached eof is already tracked in the @out_eof ivar.

`pos` is only needed by minitar (as far as I can tell), but again has a
straightforward implementation - the only quirk being that bzip2 tracks
the number of decompressed bytes emitted across 2 uints, one
representing the high part of a 64 bit value and one representing the
low part, hence the bitshifting to combine them.

With these two methods in place it's straightforward to write a wrapper
for Bzip2::FFI::Reader for use with minitar or Gem::Package::TarReader:

    class BzipReader
      extend Forwardable

      attr_reader :io, :reader
      def initialize(io)
        @io = io
        @start_pos = io.pos
        @reader = Bzip2::FFI::Reader.new(io)
      end

      def rewind
        io.pos = @start_pos
        @reader = Bzip2::FFI::Reader.new(io)
      end

      def_delegators :reader,
                     :close,
                     :eof?,
                     :pos,
                     :read
    end

I've not investigated whether Bzip2::FFI::Writer is compatible with
either of the tar archivers mentioned.